+++ point.h
@@ -3,9 +3,9 @@
 
 struct Point
 {
-    double x, y, z;
+    int x, y, z;
     Point();
-    Point(double x, double y, double z);
+    Point(int x, int y, int z);
     friend std::istream &operator>>(std::istream &input, Point &p);
     friend std::ostream &operator<<(std::ostream &output, Point const &p);
     Point &operator-=(Point &p);
